# PERF: Fixed regression in Series(index=idx) constructor