From https://github.com/pandas-dev/pandas/pull/18496/

Special cases empty series construction, since the reindex is not necessary.

xref https://github.com/pandas-dev/pandas/issues/18532#issuecomment-385190350

Setup

```python
data = None
idx = date_range(start=datetime(2015, 10, 26),
                              end=datetime(2016, 1, 1),
                              freq='50s')
dict_data = dict(zip(idx, range(len(idx))))
data = None if data is None else dict_data
```

benchmark

```
%timeit out = Series(data, index=idx)
```

HEAD:

```
113 µs ± 6.22 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

0.21.0

```
96.5 µs ± 6.05 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Master

```
320 ms ± 11.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```