@TomAugspurger 2018-04-30
Good call on the init_dict. For some reason I thought it had to be in the main Series.__init__, since _init_dict returns a BlockManager & index, but this seems to work.

```
In [4]: %timeit out = Series(data, index=idx)
140 µs ± 1.24 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Shouldn't need a whatsnew since https://github.com/pandas-dev/pandas/pull/18496 was in 0.23. I think that was the original slowdown.

And this was caught by the `series_methods.SeriesConstructor.time_constructor(None)` asv, so we're good there (that's the benchmark I've been running).