@TomAugspurger 2018-04-30
Does d8b1312 change your opinion on putting this stuff in `_init_dict`? It's getting borderline too messy now.

`Series._init_dict` sorts the index (keys). This would break

```python
Series(index=['b', 'a'])
```

Since the user input isn't actually dict-like, it's just going down _init_dict.