@toobaz 2018-04-30
I might be wrong (can't test now, sorry), but a simpler solution could be to replace

https://github.com/pandas-dev/pandas/blob/4afc7563895830d224ac949f571edef2f069c314/pandas/core/series.py#L180

with

``` python
if data in [None, {}]:
    data = []
```