@TomAugspurger 2018-04-30
That won't quiet work since `data` may not be hashable, but I like the
thought. Will try a variant.

On Mon, Apr 30, 2018 at 8:34 AM, Pietro Battiston <notifications@github.com>
wrote:

> I might be wrong (can't test now, sorry), but a simpler solution could be
> to replace
>
> https://github.com/pandas-dev/pandas/blob/4afc7563895830d224ac949f571ede
> f2f069c314/pandas/core/series.py#L180
>
> with
>
> if data in [None, {}]:
>     data = []
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/20865#issuecomment-385399351>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABQHIk_2hzXwI065z48-lX8D9FfUQJWXks5ttxLQgaJpZM4Trxgq>
> .
>