@TomAugspurger 2018-04-30
Merging in 2 hours if there are no further comments. I think we're just choosing between the two least-bad outcomes here.