@jreback 2018-04-30
lgtm. thanks (I believe you said this was after 0.22 so no whatsnew needed)